# docs/api/quant-radar-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Surge Screener API
  version: 1.23.0-draft
  summary: Loopback-only public reads plus protected private operator state.
  description: |
    Contract for the current Phase 1 fixed-artifact reads plus bounded Phase 2
    reads for IV history, the institutional fund catalog, the manually maintained
    AI updates feed, the manually maintained schedule registry, a strict
    Options Flow and candidate feed projections, the strict Crypto Universe,
    Social Intelligence, Theme Flow, Money Flow, Sector Rotation, Daily Summary,
    Playbook Validation, Continuation Validation, COT report catalog/detail, and
    bounded Market Thesis validation/regime-history, Oversold Reversal
    validation, Reversal Radar validation, Knowledge Graph, Watchlist theme
    taxonomy, sector-to-theme drill, and Influencer roster public projections.
    Expected
    artifact failures are represented as HTTP 200 unavailable envelopes so a
    missing or half-written scheduled output cannot crash the frontend. Invalid
    request parameters use sanitized HTTP 422 problem details. Unexpected server
    defects remain HTTP 500.

    Public artifact routes intentionally have no authentication scheme because
    the server is restricted to 127.0.0.1 and exposes bounded non-sensitive
    projections. The private Industry Roles review-board routes additionally
    require a high-entropy Streamlit-to-FastAPI bearer credential. Mutations
...
servers:
  - url: http://127.0.0.1:8000
...
security: []
...
paths:
...
  /api/v1/reports/continuation-validation/latest:
    get:
      tags: [Reports]
      summary: Read the latest Continuation Validation presentation
      description: >-
        Reads one fixed server-owned validation artifact and returns the bounded
        maturity summary and displayed ticker outcomes. Private blocked reasons,
        producer notes, feature corpora, provider details, and paths remain
        server-side.
      operationId: getLatestContinuationValidation
      responses:
        "200":
          $ref: "#/components/responses/ContinuationValidationResponse"
        "500":
          $ref: "#/components/responses/InternalProblem"
...
components:
  securitySchemes:
    InternalServiceBearer:
      type: http
      scheme: bearer
      description: |
        High-entropy Streamlit-to-FastAPI service credential for protected
        loopback routes; not a human login or provider token.
  headers:
    NoStore:
      description: Prevent browsers and intermediaries from retaining artifact state.
      schema:
        type: string
        const: no-store
...
  responses:
...
    ContinuationValidationResponse:
      description: Strict latest Continuation Validation presentation state.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ContinuationValidationEnvelope"
          examples:
            available:
              summary: An accumulating Continuation Validation report
              value:
                available: true
                reason: ok
                data:
                  generated_at: "2026-08-05T01:02:03+00:00"
                  status: accumulating
                  resolved: 1
                  min_resolved: 2
                  summary:
                    strong_continuation: 1
                    normal_continuation: 0
                    failed_breakout: 0
                    unresolved: 0
                    rows_total: 1
                    resolved: 1
                    min_resolved: 2
                  rows:
                    - ticker: NVDA
                      setup_date: "2026-01-05"
                      surge_start: "2026-01-02"
                      thresholds_hit: ["+30%/20d"]
                      magnitude_pct: 42.0
                      candidate_causes: [technical_volume_expansion]
                      cause_certainty: candidate_only
                      measurement_source: daily_bars
                      resolved_30d: true
                      fwd_30d_return: 0.18
                      fwd_30d_max_drawdown: -0.05
                      resolved_60d: true
                      fwd_60d_return: 0.34
                      fwd_60d_max_drawdown: -0.13
                      continuation_label: strong_continuation
                      primary_horizon: 30d
                      trade_value: high
                meta:
                  sourceId: reports.continuation-validation.latest
                  asOf: null
                  generatedAt: "2026-08-05T01:02:03+00:00"
            blocked:
              summary: A blocked report without its private diagnostic reason
              value:
                available: true
                reason: ok
                data:
                  generated_at: "2026-08-05T01:02:03+00:00"
                  status: blocked
                  resolved: 0
                  min_resolved: 2
                  summary:
                    strong_continuation: 0
                    normal_continuation: 0
                    failed_breakout: 0
                    unresolved: 0
                    rows_total: 0
                    resolved: 0
                    min_resolved: 2
                  rows: []
                meta:
                  sourceId: reports.continuation-validation.latest
                  asOf: null
                  generatedAt: "2026-08-05T01:02:03+00:00"
            unavailable:
              summary: The fixed Continuation Validation artifact is missing
              value:
                available: false
                reason: missing
                data: null
                meta:
                  sourceId: reports.continuation-validation.latest
                  asOf: null
                  generatedAt: null
...
    InternalProblem:
      description: Unexpected programming or server failure.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: about:blank
            title: Internal Server Error
            status: 500
  schemas:
...
    AvailableArtifactBase:
      type: object
      required: [available, reason, meta]
      properties:
        available:
          type: boolean
          const: true
        reason:
          type: string
          const: ok
        meta:
          $ref: "#/components/schemas/ArtifactMeta"
...
    AvailableContinuationValidation:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/AvailableArtifactBase"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/ContinuationValidationData"
...
    UnavailableArtifact:
      type: object
      additionalProperties: false
      required: [available, reason, data, meta]
      properties:
        available:
          type: boolean
          const: false
        reason:
          $ref: "#/components/schemas/UnavailableReason"
        data:
          type: "null"
        meta:
          $ref: "#/components/schemas/ArtifactMeta"
    UnavailableReason:
      type: string
      enum:
        - missing
        - invalid_json
        - invalid_shape
        - unreadable
...
    ContinuationValidationEnvelope:
      oneOf:
        - $ref: "#/components/schemas/AvailableContinuationValidation"
        - $ref: "#/components/schemas/UnavailableArtifact"
...
    ContinuationValidationData:
      type: object
      additionalProperties: false
      required: [generated_at, status, resolved, min_resolved, summary, rows]
      properties:
        generated_at:
          type: string
          format: date-time
          minLength: 20
          maxLength: 32
          pattern: '^\d{4}-\d{2}-\d{2}[Tt](?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d{1,6})?(?:[Zz]|[+-](?:[01]\d|2[0-3]):[0-5]\d)(?![\s\S])'
        status:
          type: string
          enum: [blocked, accumulating, ready]
        resolved:
          type: integer
          minimum: 0
          maximum: 5000
        min_resolved:
          type: integer
          minimum: 1
          maximum: 1000000
        summary:
          $ref: "#/components/schemas/ContinuationValidationSummary"
        rows:
          type: array
          maxItems: 1000
          items:
            $ref: "#/components/schemas/ContinuationValidationRow"
    ContinuationValidationSummary:
      type: object
      additionalProperties: false
      required:
        - strong_continuation
        - normal_continuation
        - failed_breakout
        - unresolved
        - rows_total
        - resolved
        - min_resolved
      properties:
        strong_continuation:
          type: integer
          minimum: 0
          maximum: 5000
        normal_continuation:
          type: integer
          minimum: 0
          maximum: 5000
        failed_breakout:
          type: integer
          minimum: 0
          maximum: 5000
        unresolved:
          type: integer
          minimum: 0
          maximum: 5000
        rows_total:
          type: integer
          minimum: 0
          maximum: 5000
        resolved:
          type: integer
          minimum: 0
          maximum: 5000
        min_resolved:
          type: integer
          minimum: 1
          maximum: 1000000
    ContinuationValidationRow:
      type: object
      additionalProperties: false
      required:
        - ticker
        - setup_date
        - surge_start
        - thresholds_hit
        - magnitude_pct
        - candidate_causes
        - cause_certainty
        - measurement_source
        - resolved_30d
        - fwd_30d_return
        - fwd_30d_max_drawdown
        - resolved_60d
        - fwd_60d_return
        - fwd_60d_max_drawdown
        - continuation_label
        - primary_horizon
        - trade_value
      properties:
        ticker:
          type: string
          minLength: 1
          maxLength: 15
          pattern: '^[A-Z0-9]+(?:[.-][A-Z0-9]+)*(?![\s\S])'
        setup_date:
          type: string
          format: date
          pattern: '^\d{4}-\d{2}-\d{2}(?![\s\S])'
        surge_start:
          anyOf:
            - type: string
              format: date
              pattern: '^\d{4}-\d{2}-\d{2}(?![\s\S])'
            - type: "null"
        thresholds_hit:
          type: array
          maxItems: 100
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 100
            pattern: '^\S(?:[\s\S]*\S)?(?![\s\S])'
        magnitude_pct:
          anyOf:
            - type: number
              minimum: -100
              maximum: 1000000000
            - type: "null"
        candidate_causes:
          type: array
          minItems: 1
          maxItems: 7
          uniqueItems: true
          items:
            type: string
            enum:
              - technical_volume_expansion
              - relative_strength_leadership
              - technical_compression_base
              - technical_momentum_confirm
              - trend_quality
              - market_regime_support
              - unknown
        cause_certainty:
          type: string
          const: candidate_only
        measurement_source:
          type: string
          const: daily_bars
        resolved_30d:
          type: boolean
        fwd_30d_return:
          $ref: "#/components/schemas/ContinuationOptionalReturn"
        fwd_30d_max_drawdown:
          $ref: "#/components/schemas/ContinuationOptionalReturn"
        resolved_60d:
          type: boolean
        fwd_60d_return:
          $ref: "#/components/schemas/ContinuationOptionalReturn"
        fwd_60d_max_drawdown:
          $ref: "#/components/schemas/ContinuationOptionalReturn"
        continuation_label:
          type: string
          enum: [strong_continuation, normal_continuation, failed_breakout, unresolved]
        primary_horizon:
          anyOf:
            - type: string
              enum: [30d, 60d]
            - type: "null"
        trade_value:
          type: string
          enum: [high, medium, low, unknown]
    ContinuationOptionalReturn:
      anyOf:
        - type: number
          minimum: -1
          maximum: 1000000000
        - type: "null"
...
    ArtifactMeta:
      type: object
      additionalProperties: false
      required: [sourceId, asOf, generatedAt]
      properties:
        sourceId:
          type: string
          description: Stable registry identifier; never a filesystem path.
          pattern: "^[a-z0-9]+(?:[.-][a-z0-9]+)*$"
        asOf:
          type: [string, "null"]
          format: date
        generatedAt:
          type: [string, "null"]
          format: date-time
    Problem:
      type: object
      description: RFC 9457 problem details for unexpected server failures.
      additionalProperties: true
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference